refactor: updated names of actions component to be same as names used in API
