+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/LinkAction/index.ts
@@ -1,0 +1,1 @@
+export { LinkAction } from './LinkAction'
