+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateAction/index.ts
@@ -1,0 +1,1 @@
+export { NavigateAction } from './NavigateAction'
