+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateToBlockAction/index.ts
@@ -1,0 +1,1 @@
+export { NavigateToBlockAction } from './NavigateToBlockAction'
