+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/NavigateToJourneyAction/index.ts
@@ -1,0 +1,1 @@
+export { NavigateToJourneyAction } from './NavigateToJourneyAction'
