+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/Action/index.ts
@@ -1,0 +1,1 @@
+export { Action } from './Action'
