--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/ActionProperties/NavigateJourney/index.ts
@@ -1,1 +1,0 @@
-export { NavigateJourney } from './NavigateJourney'
