--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/ActionProperties/NavigateLink/index.ts
@@ -1,1 +1,0 @@
-export { NavigateLink } from './NavigateLink'
