--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/ActionProperties/NavigateNext/index.ts
@@ -1,1 +1,0 @@
-export { NavigateNext } from './NavigateNext'
