--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/ActionProperties/NavigateStep/index.ts
@@ -1,1 +1,0 @@
-export { NavigateStep } from './NavigateStep'
