--- apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/ActionProperties/index.ts
@@ -1,1 +1,0 @@
-export { ActionProperties } from './ActionProperties'
